My_AI_Employee/pyproject.toml
[project]
name = "my-ai-employee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Syedaashnaghazanfar", email = "nutellacookiesss@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "python-frontmatter>=1.1.0",
    "watchdog>=6.0.0",
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.158.0",
    "fastmcp>=0.3.0",
    "playwright>=1.49.0",
    "requests>=2.32.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
